requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==5.0.6
django-dotenv==1.4.2
django-filter==24.2
django-jazzmin==3.0.0
django-sass-processor==1.2.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
drf-yasg==1.21.7
fonttools==4.53.0
html5lib==1.1
idna==3.7
inflection==0.5.1
libsass==0.22.0
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
python-barcode==0.15.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.3
tinycss2==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
weasyprint==62.2
webencodings==0.5.1
zopfli==0.2.3
